create a guide for the assistants' integration (v2); added req2test to assistants folder
