+++ README.md
@@ -1,2 +1,5 @@
 # main
 This is the main repository for DS 2024-25. It should contain all relevant knowledge about the products under development, architecture and design decisions, and core infrastructure.
+
+## Integrating your Assistant
+Please follow this [guide](/avengers/frontend/webapp/src/app/assistants/README) to integrate your assistant into **AI4SD - Avengers**.
