+++ avengers/frontend/webapp/src/app/assistants/example/components/assistant.js
@@ -1,0 +1,9 @@
+import styles from "@/app/page.module.css";
+export default function Assistant() {
+    return (
+        <div className={styles.assistantInteraction}>
+            <h1 className="text-3xl font-bold text-indigo-600">YOUR ASSISTANT</h1>
+            <p>Example component for YOUR ASSISTANT</p>
+        </div>
+    )
+}
