+++ avengers/frontend/webapp/src/app/assistants/req2test/assets/logoName.png
