+++ avengers/frontend/webapp/src/app/assistants/req2test/components/assistant.js
@@ -1,7 +1,7 @@
 import styles from "@/app/page.module.css";
 import { Chatbot } from './Chat';
 
-export default function Req2Test() {
+export default function Assistant() {
     return (
         <div className={styles.assistantInteraction}>
             <Chatbot />
